+++ components/ui/Tooltip/index.tsx
@@ -1,40 +1,34 @@
 import * as React from "react";
-import * as Tooltip from "@radix-ui/react-tooltip";
+import * as Popover from "@radix-ui/react-popover";
 import classNames from "classnames";
 
-interface LTooltipProps
+interface LPopoverProps
   extends Omit<React.HTMLAttributes<HTMLElement>, "title"> {
-  delayDuration?: number;
   title?: React.ReactNode;
 }
 
-const LTooltip: React.FC<LTooltipProps> = ({
-  children,
-  delayDuration = 500,
-  title,
-}) => {
+const LPopover: React.FC<LPopoverProps> = ({ children, title }) => {
   return (
-    <Tooltip.Provider delayDuration={delayDuration}>
-      <Tooltip.Root>
-        <Tooltip.Trigger>{children}</Tooltip.Trigger>
-        <Tooltip.Portal>
-          <Tooltip.Content
-            align="start"
-            className={classNames(
-              "z-[2000] py-2 px-4 text-sm rounded-md shadow-md max-w-xs break-words",
-              "data-[state=delayed-open]:animate-fadeInUp data-[state=closed]:animate-fadeOut",
-              "bg-neutral-800 text-white/90",
-              "dark:bg-neutral-700 dark:text-white/90"
-            )}
-            sideOffset={5}
-          >
-            {title}
-            <Tooltip.Arrow className="fill-neutral-800 dark:fill-neutral-700" />
-          </Tooltip.Content>
-        </Tooltip.Portal>
-      </Tooltip.Root>
-    </Tooltip.Provider>
+    <Popover.Root>
+      <Popover.Trigger>{children}</Popover.Trigger>
+      <Popover.Portal>
+        <Popover.Content
+          side="top"
+          align="start"
+          className={classNames(
+            "z-[2000] py-2 px-4 text-sm rounded-md shadow-md max-w-popover md:max-w-xs break-words",
+            "data-[state=open]:animate-fadeInUp data-[state=closed]:animate-fadeOut",
+            "bg-neutral-800 text-white/90",
+            "dark:bg-neutral-700 dark:text-white/90"
+          )}
+          sideOffset={5}
+        >
+          {title}
+          <Popover.Arrow className="fill-neutral-800 dark:fill-neutral-700" />
+        </Popover.Content>
+      </Popover.Portal>
+    </Popover.Root>
   );
 };
 
-export default LTooltip;
+export default LPopover;