+++ CHANGE_LOG.md
@@ -1,5 +1,14 @@
 # L-GPT Change Log
 
+## v0.1.2
+
+> 2023-04-26
+
+### Add
+
+- Add the ability to configure max_tokens
+- Add the ability to select a model
+
 ## v0.1.1
 
 > 2023-04-24